requirements.txt
Pyrogram==2.0.106
tgcrypto==1.2.5
requests==2.31.0
pyromod==1.5
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
mutagen==1.46.0
pyaes==1.6.1
ffmpeg
mediainfo
PySocks==1.7.1
python-dotenv==1.0.0
cloudscraper
beautifulsoup4==4.9.3
bs4==0.0.1
aiohttp==3.8.4
aiofiles
yt-dlp==2023.6.22
pycryptodome==3.18.0
PySocks==1.7.1
websockets==9.1
psutil
async-lru==2.0.2
helper
pyaes==1.6.1
soupsieve==2.4.1
urllib3==2.0.3
websockets==11.0.3
py-tgcalls==0.6.0
motor
pytz
umongo==3.1.0
speedtest-cli